fix(game.js): fix star collection bug
This commit fixes a bug that caused the score to jump by too much when a star was collected. The issue was caused by the logic for removing a star being on the server, so that when the ship ovrlaps a star it takes some time for the star to disappear. This resulted in the overlap method emitting too many 'starCollected' events to the server. The solution was to destroy the star on the client first.