fix(all): fix duplicate ships bug
This commit fixes an issue that causes new ship instances to be created when the socket connection auto-disconnects and reconnects. The solution was to use cookies to create a user session, using the express-session package, and share it with the socket connection using the express-socket.io-session package. The current implementation doesn't persist the express-socket.io-session cookie through reloads or closing the browser window. To get this functionality, simply pass { autoSave:true } into the second argument of sharedsession().